ATS Analyzer/ATS Analyzer/requirements.txt
Flask==2.3.3
pdfplumber==0.9.0
python-docx==0.8.11
nltk==3.8.1
spacy==3.6.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7